feat: implement DenylistTable and refactor FallbackDenylistService to use it
